Split tests of elementwise functions into separate files
